+++ app/src/main/java/br/com/bootcamp/magicgamecs/features/favorites/FavoritesFragment.kt
@@ -1,0 +1,21 @@
+package br.com.bootcamp.magicgamecs.features.favorites
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import br.com.bootcamp.magicgamecs.R
+
+class FavoritesFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_favorites, container, false)
+    }
+
+}
