+++ app/src/main/res/layout/fragment_favorites.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".features.favorites.FavoritesFragment">
+
+    <!-- TODO: Update blank fragment layout -->
+
+
+</FrameLayout>
