+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_collection"
+        android:title="@string/general_list"
+        android:icon="@drawable/cards"/>
+
+    <item
+        android:id="@+id/navigation_favorites"
+        android:title="@string/favorites"
+        android:icon="@drawable/heart"/>
+
+</menu>
