+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@+id/navigation_collection">
+
+    <fragment
+        android:id="@+id/navigation_collection"
+        android:name="br.com.bootcamp.magicgamecs.features.home.CollectionFragment"
+        tools:layout="@layout/fragment_collection"/>
+
+    <fragment
+        android:id="@+id/navigation_favorites"
+        android:name="br.com.bootcamp.magicgamecs.features.favorites.FavoritesFragment"
+        tools:layout="@layout/fragment_favorites"/>
+
+
+</navigation>
